minor-update functions
