+++ templates/modules/DB/Frame_DBFrame.py
@@ -5,7 +5,7 @@ __date__ = '$ 18/oct./2023  at 10:25 $'
 import ttkbootstrap as ttk
 
 from static.extensions import windows_names_db_frame
-from templates.Functions_AuxFiles import get_image_side_menu
+from templates.Functions_AuxFilesGUI import get_image_side_menu
 from templates.Funtions_Utils import create_button_side_menu
 from templates.modules.DB.SubFrame_Chats import ChatsFrame
 from templates.modules.DB.SubFrame_Customer import CustomersFrame
