+++ templates/modules/RRHH/SubFrame_FichajesManual.py
@@ -11,10 +11,11 @@ from ttkbootstrap.tableview import Tableview
 
 from static.extensions import files_fichaje_path, patterns_files_fichaje, cache_file_emp_fichaje
 from templates.Functions_AuxFiles import get_events_op_date
-from templates.Functions_Files import create_spinboxes_time, get_info_f_file_name, get_info_bitacora, \
+from templates.Functions_Files import get_info_f_file_name, get_info_bitacora, \
     get_info_t_file_name, check_fichajes_files_in_directory, extract_fichajes_file, \
     check_names_employees_in_cache, extract_data_file_contracts, generate_table_from_dict_contracts
-from templates.Funtions_Utils import create_var_none, create_Combobox, create_label, create_stringvar, update_stringvars
+from templates.Funtions_Utils import create_var_none, create_Combobox, create_label, create_stringvar, \
+    update_stringvars, create_spinboxes_time
 from templates.modules.Misc.Frame_CollapsingFrame import CollapsingFrame
 from templates.modules.Misc.SubFrame_Plots import FramePlot
